requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
sqlparse==0.5.0
